+++ dlls/globals.cpp
@@ -28,9 +28,9 @@
 #include "cbase.h"
 #include "soundent.h"
 
-DLL_GLOBAL ULONG		g_ulFrameCount;
-DLL_GLOBAL ULONG		g_ulModelIndexEyes;
-DLL_GLOBAL ULONG		g_ulModelIndexPlayer;
+DLL_GLOBAL unsigned int g_ulFrameCount;
+DLL_GLOBAL unsigned int g_ulModelIndexEyes;
+DLL_GLOBAL unsigned int g_ulModelIndexPlayer;
 DLL_GLOBAL Vector		g_vecAttackDir;
 DLL_GLOBAL int			g_iSkillLevel;
 DLL_GLOBAL int			gDisplayTitle;
